Switching to Gtihub
We started to work on this project on site .Thus we didn't use Github until now . In the initial commit , you can find the first version of our project : The robot detects a start signal and  searches the ball by rotating and using tof sensor. Once detected , the e-puck moves toward the ball , hits it and stops . If it detects the goal signal , it starts a celebration (leds + rotation).